+++ ydb/topic.py
@@ -11,6 +11,7 @@ __all__ = [
     "TopicMeteringMode",
     "TopicReader",
     "TopicReaderAsyncIO",
+    "TopicReaderSelector",
     "TopicReaderSettings",
     "TopicStatWindow",
     "TopicWriter",
@@ -30,6 +31,7 @@ from . import driver
 
 from ._topic_reader.topic_reader import (
     PublicReaderSettings as TopicReaderSettings,
+    PublicTopicSelector as TopicReaderSelector,
 )
 
 from ._topic_reader.topic_reader_sync import TopicReaderSync as TopicReader
@@ -151,7 +153,7 @@ class TopicClientAsyncIO:
 
     def reader(
         self,
-        topic: str,
+        topic: Union[str, TopicReaderSelector, List[Union[str, TopicReaderSelector]]],
         consumer: str,
         buffer_size_bytes: int = 50 * 1024 * 1024,
         # decoders: map[codec_code] func(encoded_bytes)->decoded_bytes
@@ -300,7 +302,7 @@ class TopicClient:
 
     def reader(
         self,
-        topic: str,
+        topic: Union[str, TopicReaderSelector, List[Union[str, TopicReaderSelector]]],
         consumer: str,
         buffer_size_bytes: int = 50 * 1024 * 1024,
         # decoders: map[codec_code] func(encoded_bytes)->decoded_bytes